fix: date is shared & ScheduleType is schedule model
