+++ apps/web/src/components/ui/Schedule/ScheduleList.tsx
@@ -1,9 +1,9 @@
 import { Button } from '@heroui/react';
 import { Suspense, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
+import { ScheduleType } from '@entities/schedule/model/ScheduleType';
 import { useGetScheduleByDate } from '@hooks/apollo/Schedule';
 import { useScheduleActions } from '@hooks/useScheduleMenu';
-import { ScheduleType } from '@shared/lib/utils';
 import StateRender from '@shared/ui/StateRender';
 import MenuableAccordion from '@ui/CustomComponent/MenuableAccordion';
 import ScheduleDisplay from './ScheduleDisplay';
