+++ apps/web/src/entities/schedule/model/ScheduleType.ts
@@ -1,0 +1,9 @@
+import { enumify } from '@shared/lib/utils';
+
+export const ScheduleType = enumify({
+  BREAK: 'BREAK',
+  SCHEDULED: 'SCHEDULED',
+  STARTED: 'STARTED',
+  PAUSED: 'PAUSED',
+  FINISH: 'FINISH',
+})
