+++ apps/web/src/hooks/apollo/Schedule/index.ts
@@ -1,5 +1,5 @@
-import DateService from '@entities/Calender/model/DateService'
-import { ScheduleType } from '@shared/lib/utils'
+import { ScheduleType } from '@entities/schedule/model/ScheduleType'
+import { DateService } from '@shared/ui/Calender'
 import useCloneSchedule from './useCloneSchedule'
 import useCloneScheduleFromPreset from './useCloneScheduleFromPreset'
 import useCreateSchedule from './useCreateSchedule'
