+++ apps/web/src/hooks/apollo/mixed/useCreateScheduleWithExercisePlans.ts
@@ -1,6 +1,6 @@
+import { ScheduleType } from '@entities/schedule/model/ScheduleType';
 import { useCreateExerciseBySchedule } from '@hooks/apollo/Exercise'
 import { useCreateSchedule } from '@hooks/apollo/Schedule'
-import { ScheduleType } from '@shared/lib/utils'
 
 export default function useCreateScheduleWithExercisePlans() {
   const [createSchedule] = useCreateSchedule()
