+++ apps/web/src/hooks/useScheduleMenu/useScheduleActions.ts
@@ -1,9 +1,9 @@
 import type { NavigateOptions } from 'react-router-dom';
 import { useCallback } from 'react'
 import { useNavigate } from 'react-router-dom'
+import { ScheduleType } from '@entities/schedule/model/ScheduleType';
 import { useAlert } from '@globalUi/Alert';
 import { useCloneSchedule, useCreateSchedule, useDeleteSchedule } from '@hooks/apollo/Schedule'
-import { ScheduleType } from '@shared/lib/utils'
 
 export default function useScheduleActions() {
   const navigate = useNavigate()
