+++ apps/web/src/hooks/useScheduleMenu/useScheduleHeaderMenu.tsx
@@ -2,11 +2,11 @@ import type { Schedule } from 'fitness-struct'
 import { useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigate } from 'react-router-dom'
+import { ScheduleType } from '@entities/schedule/model/ScheduleType';
 import { useAlert } from '@globalUi/Alert';
 import { useCopyExercisePresetFromSchedule } from '@hooks/apollo/ExercisePreset'
 import useScheduleMenu from '@hooks/useScheduleMenu'
 import { useHeaderMenuHandler } from '@shared/hooks/header';
-import { ScheduleType } from '@shared/lib/utils'
 
 export default function useScheduleHeaderMenu(scheduleInfo?: Schedule.Schedule): [
   boolean,
