+++ apps/web/src/pages/:selectDate/workout/:id.tsx
@@ -4,13 +4,14 @@ import { useAnimationFrame } from 'framer-motion'
 import { startTransition, Suspense, useCallback, useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigate, useParams } from 'react-router-dom'
+import { ScheduleType } from '@entities/schedule/model/ScheduleType'
 import { useAlert } from '@globalUi/Alert';
 import { useGetScheduleById, useUpdateSchedule } from '@hooks/apollo/Schedule'
 import usePageTracker from '@hooks/usePageTracker'
 import { useScheduleHeaderMenu } from '@hooks/useScheduleMenu'
 import { BottomNaviArea } from '@shared/hooks/bottomNavi'
 import { useHeaderHandler } from '@shared/hooks/header'
-import { dayjs, ScheduleType } from '@shared/lib/utils'
+import { dayjs } from '@shared/lib/utils'
 import ExerciseDataList from '@ui/ExerciseData/ExerciseDataList'
 import PresetNameInputDialog from '@ui/Preset/PresetNameInputDialog'
