+++ apps/web/src/pages/Main.tsx
@@ -2,12 +2,12 @@ import { Button, ScrollShadow } from '@heroui/react'
 import { Suspense, useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigate } from 'react-router-dom'
-import DateService from '@entities/Calender/model/DateService'
 import { useGetScheduleByDate } from '@hooks/apollo/Schedule'
 import usePageTracker from '@hooks/usePageTracker'
 import { useScheduleActions } from '@hooks/useScheduleMenu'
 import { useBottomNavi } from '@shared/hooks/bottomNavi'
 import { useHeaderHandler } from '@shared/hooks/header'
+import { DateService } from '@shared/ui/Calender'
 import StateRender from '@shared/ui/StateRender'
 import MenuableAccordion from '@ui/CustomComponent/MenuableAccordion'
 import ScheduleDisplay from '@ui/Schedule/ScheduleDisplay'
