+++ apps/web/src/shared/ui/Calender/index.ts
@@ -1,0 +1,3 @@
+export { default } from './ui/templates/Calender'
+export type { DateValue, DateRange } from './types'
+export { default as DateService } from './model/DateService'
