+++ apps/web/src/shared/ui/Calender/ui/atoms/DateButton.tsx
@@ -1,22 +1,19 @@
 import type { DateButtonProps } from '../types';
 import { Button } from '@heroui/react';
-import { calenderColor } from '@entities/Calender/lib/color';
-import DateService from '@entities/Calender/model/DateService';
+import { DateService } from '@shared/ui/Calender';
 
 export default function DateButton({
-  value,
   display,
   dateRange,
   onChange,
-  statesByDate = []
+  color
 }: DateButtonProps) {
-  const colorClass = calenderColor(DateService.isEqual(display, value), statesByDate)
   const inRange = DateService.isDateInRange(display, dateRange)
   return <Button
     isDisabled={!inRange}
     variant='bordered'
     onPress={() => onChange(display)}
-    className={['flex-1 px-0 min-w-[40px]', colorClass].join(' ')}
+    className={['flex-1 px-0 min-w-[40px]', color].join(' ')}
     radius='full'
   >
     {display.date}
