+++ apps/web/src/shared/ui/Calender/ui/types.ts
@@ -1,5 +1,4 @@
 import type { DateValue, DateRange } from '../types';
-import type { Schedule } from 'fitness-struct';
 import type { ReactNode } from 'react';
 
 export type CalendarMode = 'date' | 'month' | 'year';
@@ -11,12 +10,12 @@ export interface CalenderBaseProps {
 }
 
 export interface DateCalenderStates {
-  statesByMonth?: Schedule.IType[][]
+  colorByDate?: string[]
 }
 
 export interface DateButtonProps extends CalenderBaseProps {
   display: DateValue
-  statesByDate?: Schedule.IType[]
+  color?: string
 }
 
 export interface DateGridProps extends CalenderBaseProps, DateCalenderStates {
